Changed SciterColor constructor implementation to use Byte (from Int32). Removed `SciterColor.Invalid` as it's not something that will ever be valid. Updated tests.

+++ Shared/SciterCore.Shared/Extensions/SciterGraphicsExtensions.cs
@@ -174,19 +174,19 @@ namespace SciterCore
             return graphics;
         }
         
-        public static SciterGraphics SetLineColor(this SciterGraphics graphics, int r, int g, int b)
+        public static SciterGraphics SetLineColor(this SciterGraphics graphics, byte r, byte g, byte b)
         {
             graphics?.SetLineColor(r: r, g: g, b: b, alpha: 1f);
             return graphics;
         }
         
-        public static SciterGraphics SetLineColor(this SciterGraphics graphics, int r, int g, int b, float alpha)
+        public static SciterGraphics SetLineColor(this SciterGraphics graphics, byte r, byte g, byte b, float alpha)
         {
             graphics?.SetLineColor(lineColor: new SciterColor(r: r, g: g, b: b, alpha: alpha));
             return graphics;
         }
 
-        public static SciterGraphics SetLineColor(this SciterGraphics graphics, int r, int g, int b, int alpha)
+        public static SciterGraphics SetLineColor(this SciterGraphics graphics, byte r, byte g, byte b, byte alpha)
         {
             graphics?.SetLineColor(lineColor: new SciterColor(r: r, g: g, b: b, alpha: alpha));
             return graphics;
@@ -197,17 +197,17 @@ namespace SciterCore
             return graphics?.TrySetLineColorInternal(lineColor: lineColor) == true;
         }
 
-        public static bool TrySetLineColor(this SciterGraphics graphics, int r, int g, int b)
+        public static bool TrySetLineColor(this SciterGraphics graphics, byte r, byte g, byte b)
         {
             return graphics?.TrySetLineColor(r: r, g: g, b: b, alpha: 1f) == true;
         }
 
-        public static bool TrySetLineColor(this SciterGraphics graphics, int r, int g, int b, float alpha = 1f)
+        public static bool TrySetLineColor(this SciterGraphics graphics, byte r, byte g, byte b, float alpha = 1f)
         {
             return graphics?.TrySetLineColor(lineColor: new SciterColor(r: r, g: g, b: b, alpha: alpha)) == true;
         }
 
-        public static bool TrySetLineColor(this SciterGraphics graphics, int r, int g, int b, int alpha = 255)
+        public static bool TrySetLineColor(this SciterGraphics graphics, byte r, byte g, byte b, byte alpha = 255)
         {
             return graphics?.TrySetLineColor(lineColor: new SciterColor(r: r, g: g, b: b, alpha: alpha)) == true;
         }
@@ -218,21 +218,21 @@ namespace SciterCore
             return graphics;
         }
 
-        public static SciterGraphics SetFillColor(this SciterGraphics graphics, int r, int g, int b)
+        public static SciterGraphics SetFillColor(this SciterGraphics graphics, byte r, byte g, byte b)
         {
             graphics?.SetFillColor(fillColor: new SciterColor(r: r, g: g, b: b, alpha: 1f));
             return graphics;
         }
 
         // ReSharper disable once MethodOverloadWithOptionalParameter
-        public static SciterGraphics SetFillColor(this SciterGraphics graphics, int r, int g, int b, float alpha = 1f)
+        public static SciterGraphics SetFillColor(this SciterGraphics graphics, byte r, byte g, byte b, float alpha = 1f)
         {
             graphics?.SetFillColor(fillColor: new SciterColor(r: r, g: g, b: b, alpha: alpha));
             return graphics;
         }
 
         // ReSharper disable once MethodOverloadWithOptionalParameter
-        public static SciterGraphics SetFillColor(this SciterGraphics graphics, int r, int g, int b, int alpha = byte.MaxValue)
+        public static SciterGraphics SetFillColor(this SciterGraphics graphics, byte r, byte g, byte b, byte alpha = byte.MaxValue)
         {
             graphics?.SetFillColor(fillColor: new SciterColor(r: r, g: g, b: b, alpha: alpha));
             return graphics;
@@ -243,12 +243,12 @@ namespace SciterCore
             return graphics?.TrySetFillColorInternal(fillColor: fillColor) == true;
         }
         
-        public static bool TrySetFillColor(this SciterGraphics graphics, int r, int g, int b, float alpha = 1f)
+        public static bool TrySetFillColor(this SciterGraphics graphics, byte r, byte g, byte b, float alpha = 1f)
         {
             return graphics?.TrySetFillColor(fillColor: new SciterColor(r: r, g: g, b: b, alpha: alpha)) == true;
         }
 
-        public static bool TrySetFillColor(this SciterGraphics graphics, int r, int g, int b, int alpha = 255)
+        public static bool TrySetFillColor(this SciterGraphics graphics, byte r, byte g, byte b, byte alpha = 255)
         {
             return graphics?.TrySetFillColor(fillColor: new SciterColor(r: r, g: g, b: b, alpha: alpha)) == true;
         }
